requirements.txt
certifi==2021.5.30
cycler==0.10.0
h5py==2.10.0
imageai==2.1.6
Keras==2.4.3
keras-resnet==0.2.0
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.19.3
opencv-python==4.5.3.56
pandas==1.1.5
Pillow==7.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
scipy==1.4.1
six==1.16.0
wincertstore==0.2
